Refactor @foal/core for v0.40
