+++ packages/core/src/common/controller-factories/index.ts
@@ -1,0 +1,2 @@
+export { rest } from './rest.controller-factory';
+export { view } from './view.controller-factory';
