+++ packages/core/src/common/index.ts
@@ -1,0 +1,4 @@
+export * from './controller-factories';
+export * from './post-hooks';
+export * from './pre-hooks';
+export * from './services';
