+++ packages/core/src/common/post-hooks/after-that-log.post-hook.ts
@@ -1,0 +1,15 @@
+import {
+  Context,
+  postHook,
+  PostMiddleware
+} from '../../index';
+
+export function makeAfterThatLogMiddleware(message: string): PostMiddleware {
+  return function afterThatlogMiddleware(ctx: Context): void {
+    console.log(message);
+  };
+}
+
+export function afterThatLog(message: string) {
+  return postHook(makeAfterThatLogMiddleware(message));
+}
