+++ packages/core/src/common/post-hooks/index.ts
@@ -1,0 +1,1 @@
+export { afterThatLog } from './after-that-log.post-hook';
