+++ packages/core/src/common/pre-hooks/index.ts
@@ -1,0 +1,2 @@
+export { log } from './log.pre-hook';
+export { methodNotAllowed } from './method-not-allowed.pre-hook';
