+++ packages/core/src/common/pre-hooks/method-not-allowed.pre-hook.ts
@@ -1,0 +1,13 @@
+import {
+  Context,
+  MethodNotAllowedError,
+  preHook,
+} from '../../index';
+
+export async function methodNotAllowedMiddleware(ctx: Context): Promise<Context> {
+  throw new MethodNotAllowedError();
+}
+
+export function methodNotAllowed() {
+  return preHook(methodNotAllowedMiddleware);
+}
