+++ packages/core/src/common/services/index.ts
@@ -1,0 +1,9 @@
+export {
+  CreateService,
+  ReadService,
+  UpdateService,
+  DeleteService,
+  CRUDService,
+  PartialCRUDService
+} from './crud-service.interfaces';
+export { ViewService } from './view-service.interface';
