+++ packages/core/src/common/services/view-service.interface.ts
@@ -1,0 +1,5 @@
+import { ObjectType } from '../../index';
+
+export interface ViewService {
+  render(locals: ObjectType): Promise<string>|string;
+}
