--- packages/core/src/controllers/binders/index.ts
@@ -1,3 +1,0 @@
-export { ControllerBinder } from './controller.binder';
-export * from './rest';
-export * from './view';
