--- packages/core/src/controllers/binders/rest/index.ts
@@ -1,2 +1,0 @@
-export { rest } from './rest-controller.binder';
-export * from './rest-controller.interface';
