--- packages/core/src/controllers/binders/rest/rest-controller.interface.ts
@@ -1,10 +1,0 @@
-import { ObjectType } from '../../interfaces';
-
-export interface RestController {
-  create?: (data: any, query: ObjectType) => Promise<any>;
-  get?: (id: any, query: ObjectType) => Promise<any>;
-  getAll?: (query: ObjectType) => Promise<any>;
-  update?: (id: any, data: any, query: ObjectType) => Promise<any>;
-  patch?: (id: any, data: any, query: ObjectType) => Promise<any>;
-  delete?: (id: any, query: ObjectType) => Promise<any>;
-}
