--- packages/core/src/controllers/binders/view/index.ts
@@ -1,2 +1,0 @@
-export { view } from './view.binder';
-export * from './view.interface';
