--- packages/core/src/controllers/binders/view/view.interface.ts
@@ -1,5 +1,0 @@
-import { ObjectType } from '../../interfaces';
-
-export interface ViewService {
-  render(locals: ObjectType): Promise<string>|string;
-}
