--- packages/core/src/controllers/factories/index.ts
@@ -1,2 +1,0 @@
-export { postHook } from './post-hook';
-export { preHook } from './pre-hook';
