--- packages/core/src/controllers/index.ts
@@ -1,6 +1,0 @@
-export * from './binders';
-export * from './post-hooks';
-export * from './pre-hooks';
-export * from './factories';
-export * from './errors';
-export * from './interfaces';
