--- packages/core/src/controllers/post-hooks/combine-post-hooks.post-hook.ts
@@ -1,7 +1,0 @@
-import { Decorator } from '../interfaces';
-
-export function combinePostHooks(decorators: Decorator[]): Decorator {
-    return function combinePostHooks(target: any, methodName?: string): void {
-        decorators.reverse().forEach(decorator => decorator(target, methodName));
-    };
-}
