--- packages/core/src/controllers/post-hooks/index.ts
@@ -1,2 +1,0 @@
-export { afterThatLog } from './after-that-log.post-hook';
-export { combinePostHooks } from './combine-post-hooks.post-hook';
