--- packages/core/src/controllers/pre-hooks/combine-pre-hooks.pre-hook.ts
@@ -1,7 +1,0 @@
-import { Decorator } from '../interfaces';
-
-export function combinePreHooks(decorators: Decorator[]): Decorator {
-    return function combinePreHooks(target: any, methodName?: string): void {
-        decorators.reverse().forEach(decorator => decorator(target, methodName));
-    };
-}
