--- packages/core/src/controllers/pre-hooks/index.ts
@@ -1,3 +1,0 @@
-export { combinePreHooks } from './combine-pre-hooks.pre-hook';
-export { log } from './log.pre-hook';
-export { methodNotAllowed } from './method-not-allowed.pre-hook';
