--- packages/core/src/controllers/pre-hooks/method-not-allowed.pre-hook.ts
@@ -1,11 +1,0 @@
-import { MethodNotAllowedError } from '../errors';
-import { preHook } from '../factories';
-import { Context } from '../interfaces';
-
-export async function methodNotAllowedMiddleware(ctx: Context): Promise<Context> {
-  throw new MethodNotAllowedError();
-}
-
-export function methodNotAllowed() {
-  return preHook(methodNotAllowedMiddleware);
-}
