+++ packages/core/src/factories/index.ts
@@ -1,0 +1,3 @@
+export { postHook } from './post-hook';
+export { preHook } from './pre-hook';
+export { ControllerFactory } from './controller-factory';
