+++ packages/core/src/foal.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 
-import { Service } from './di/service-manager';
 import { Foal, FoalModule } from './foal';
+import { Service } from './service-manager';
 
 describe('Foal', () => {
 
@@ -33,15 +33,15 @@ describe('Foal', () => {
       expect(foal2.services.get(Foobar)).to.equal(foal1.services.get(Foobar));
     });
 
-    xit('should create methodBindings from the controller bindings.', () => {
+    xit('should create lowLevelRoutes from the controller bindings.', () => {
 
     });
 
-    xit('should create methodBindings from the module imported.', () => {
+    xit('should create lowLevelRoutes from the module imported.', () => {
 
     });
 
-    xit('should add the module pre-hooks to the methodBindings.', () => {
+    xit('should add the module pre-hooks to the lowLevelRoutes.', () => {
       // This test relies on the two previous tests. It's not great.
 
     });
