+++ packages/core/src/index.ts
@@ -1,3 +1,8 @@
-export { ServiceManager, Service } from './di/service-manager';
-export * from './foal';
-export * from './controllers';
+export * from './common';
+export * from './factories';
+export * from './interfaces';
+export { combineHooks } from './utils';
+
+export * from './errors';
+export { Foal, FoalModule } from './foal';
+export { ServiceManager, Service } from './service-manager';
