--- packages/core/src/interfaces.ts
@@ -1,3 +1,0 @@
-export interface Type<T> {
-  new (...args: any[]): T;
-}
