+++ packages/core/src/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './contexts';
+export * from './controller-and-routes';
+export * from './middlewares';
+export * from './utils';
