+++ packages/core/src/interfaces/middlewares.ts
@@ -1,0 +1,6 @@
+import { ServiceManager } from '../service-manager';
+import { Context } from './contexts';
+
+export type Middleware = (ctx: Context) => Promise<any>|any;
+export type PreMiddleware = (ctx: Context, services: ServiceManager) => Promise<any>|any;
+export type PostMiddleware = (ctx: Context, services: ServiceManager) => Promise<any>|any;
