+++ packages/core/src/service-manager.spec.ts
